refactor(output): move FunctionRow/FileEntry to dataclasses submodule
Addresses review: extract dataclasses into complexipy/utils/dataclasses.py
and reference via submodule (dc.FunctionRow, dc.FileEntry).
